+++ common/src/main/java/net/wizards/config/LootConfig.java
@@ -15,4 +15,16 @@ public class LootConfig {
         }
     }
     public HashMap<String, List<String>> loot_tables = new HashMap<>();
+
+    public static LootConfig constrainValues(LootConfig config) {
+        if (config.item_groups != null) {
+            for (var entry: config.item_groups.entrySet()) {
+                var group = entry.getValue();
+                if (group.weight < 1) {
+                    group.weight = 1;
+                }
+            }
+        }
+        return config;
+    }
 }